#7178 [skip ci] WIP integration tests
